[#222] Prevent progress bar rendering from abort
Problem: It seems that the current progress bar rendering action is not
completely protected against exceptions.

Solution: As it is an action that runs for a short amount of time and we
really want it to not be interrputed, we can wrap it with
uninterruptibleMask. The resulting rendering loop also gets simpler with
this change.